move SplitToken to separate class
