+++ src/Microsoft.DocAsCode.Dfm/Microsoft.DocAsCode.Dfm.csproj
@@ -77,6 +77,7 @@
     <Compile Include="Rules\DfmXrefAutoLinkInlineRule.cs" />
     <Compile Include="Rules\DfmXrefShortcutInlineRule.cs" />
     <Compile Include="Rules\DfmYamlHeaderBlockRule.cs" />
+    <Compile Include="SplitToken.cs" />
     <Compile Include="Tokens\DfmFencesBlockToken.cs" />
     <Compile Include="Tokens\DfmIncludeBlockToken.cs" />
     <Compile Include="Tokens\DfmIncludeInlineToken.cs" />
